feat: finish eventType APIs, untested
