+++ api/appointmentService.yaml
@@ -96,32 +96,6 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/ArrayOfCalcomEventTypes'
-    post:
-      tags:
-        - consultant
-      summary: Add a new eventType to a consultant
-      operationId: addEventTypeToConsultant
-      parameters:
-        - name: consultantId
-          in: path
-          description: ID of onber consultant
-          required: true
-          schema:
-            type: string
-      requestBody:
-        description: EventType object that needs to be added to cal.com
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/CreateUpdateCalcomEventTypeDTO'
-        required: true
-      responses:
-        '200':
-          description: successful operation
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/CalcomEventType'
   /consultants/{consultantId}/bookings:
     get:
       tags:
@@ -312,6 +286,91 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/CalcomEventType'
+  /agencies/{agencyId}/eventTypes/{eventTypeId}/:
+    get:
+      tags:
+        - agency
+      summary: Get specific event-type by id
+      operationId: getAgencyEventTypeById
+      parameters:
+        - name: agencyId
+          in: path
+          description: ID of onber agency
+          required: true
+          schema:
+            type: integer
+            format: int64
+        - name: eventTypeId
+          in: path
+          description: ID of calcom eventType
+          required: true
+          schema:
+            type: integer
+            format: int64
+      responses:
+        '200':
+          description: successful operation
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/CalcomEventType'
+    put:
+      tags:
+        - agency
+      summary: Update an event-type with event-type object
+      operationId: updateAgencyEventType
+      parameters:
+        - name: agencyId
+          in: path
+          description: ID of onber agency
+          required: true
+          schema:
+            type: integer
+            format: int64
+        - name: eventTypeId
+          in: path
+          description: ID of calcom eventType
+          required: true
+          schema:
+            type: integer
+            format: int64
+      requestBody:
+        description: event-type object that needs to be updated
+        content:
+          application/json:
+            schema:
+              $ref: '#/components/schemas/CreateUpdateCalcomEventTypeDTO'
+        required: true
+      responses:
+        '200':
+          description: successful operation
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/CalcomEventType'
+    delete:
+      tags:
+        - agency
+      summary: Delete an event-type
+      operationId: deleteAgencyEventType
+      parameters:
+        - name: agencyId
+          in: path
+          description: ID of onber agency
+          required: true
+          schema:
+            type: integer
+            format: int64
+        - name: eventTypeId
+          in: path
+          description: ID of calcom eventType
+          required: true
+          schema:
+            type: integer
+            format: int64
+      responses:
+        '200':
+          description: successful operation
   /agencies/{agencyId}/consultants:
     get:
       tags:
@@ -374,70 +433,6 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/ArrayOfCalcomBookings'
-  /eventTypes/{eventTypeId}:
-    get:
-      tags:
-        - eventType
-      summary: Get specific event-type by id
-      operationId: getEventTypeById
-      parameters:
-        - name: eventTypeId
-          in: path
-          description: ID of calcom eventType
-          required: true
-          schema:
-            type: integer
-            format: int64
-      responses:
-        '200':
-          description: successful operation
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/CalcomEventType'
-    put:
-      tags:
-        - eventType
-      summary: Update an event-type with event-type object
-      operationId: updateEventType
-      parameters:
-        - name: eventTypeId
-          in: path
-          description: ID of calcom eventType
-          required: true
-          schema:
-            type: integer
-            format: int64
-      requestBody:
-        description: event-type object that needs to be updated
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/CreateUpdateCalcomEventTypeDTO'
-        required: true
-      responses:
-        '200':
-          description: successful operation
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/CalcomEventType'
-    delete:
-      tags:
-        - eventType
-      summary: Delete an event-type
-      operationId: deleteEventType
-      parameters:
-        - name: eventTypeId
-          in: path
-          description: ID of calcom eventType
-          required: true
-          schema:
-            type: integer
-            format: int64
-      responses:
-        '200':
-          description: successful operation
   /askers/processBooking:
     post:
       tags:
